* master: (86 commits)
  Return rawQuery
  PR feedback
  getShippingMethodsForRegion made cart aware
  Get rid of raw sql code at Cart|Order totalers
  Get rid of AU implicit in discounts
  Replace cart with its refnum and customer account id
  Replace shipping method with its price
  Replace promotion shadow object with its id
  Remove dependency on LineItemProductData from discounts
  Get rid of unused methods
  Line item product data know how to calc its price
  Merged-in changes addressed
  Merged-in changes addressed
  PR suggestions addressed
  Merge fixes
  PR suggestions addressed
  `doOrX` fixed: evaluate actions lazily
  Created a Changelog!
  Get rid of CartLineItemAdjustment dependency in discounts
  A blunder fixed
  ...

+++ phoenix-scala/phoenix/app/phoenix/models/cord/OrderShippingMethod.scala
@@ -5,7 +5,7 @@ import shapeless._
 import slick.jdbc.PostgresProfile.api._
 import core.db._
 
-case class OrderShippingMethod(id: Int = 0, cordRef: String, shippingMethodId: Int, price: Int)
+case class OrderShippingMethod(id: Int = 0, cordRef: String, shippingMethodId: Int, price: Long)
     extends FoxModel[OrderShippingMethod]
 
 object OrderShippingMethod {
@@ -18,7 +18,7 @@ class OrderShippingMethods(tag: Tag)
   def id               = column[Int]("id", O.PrimaryKey, O.AutoInc)
   def cordRef          = column[String]("cord_ref")
   def shippingMethodId = column[Int]("shipping_method_id")
-  def price            = column[Int]("price")
+  def price            = column[Long]("price")
 
   def * =
     (id, cordRef, shippingMethodId, price) <> ((OrderShippingMethod.apply _).tupled, OrderShippingMethod.unapply)
